+++ backend/interfaces/api/server/httputils/error.go
@@ -1,0 +1,10 @@
+package httputils
+
+type AppError struct {
+	Status int
+	Err    error
+}
+
+func (e AppError) Error() string {
+	return e.Err.Error()
+}
